requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
packaging==23.1
platformdirs==3.11.0
protobuf==4.21.12
PyJWT==1.7.1
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
virtualenv==20.24.6
